Refactoring to move various methods and properties to the base classes.
