+++ ResSimpy/OpenGoSim/DataModels/OpenGoSimCompletion.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from ResSimpy.Completion import Completion
-from ResSimpy.Units.AttributeMappings.BaseUnitMapping import BaseUnitMapping
+from ResSimpy.Units.AttributeMappings.CompletionUnitMapping import CompletionUnits
 
 
 @dataclass(kw_only=True)
@@ -13,6 +13,6 @@ class OpenGoSimCompletion(Completion):
         raise NotImplementedError("Not implemented for OGS yet")
 
     @property
-    def units(self) -> BaseUnitMapping:
+    def units(self) -> CompletionUnits:
         """Returns the attribute to unit map for the data object."""
         raise NotImplementedError("Not implemented for OGS yet")
