+++ ResSimpy/OpenGoSim/Enums/SimulationTypeEnum.py
